+++ packages/terra-image/CHANGELOG.md
@@ -6,6 +6,7 @@ Unreleased
 ### Added
 * Added `Object-fit-images` polyfill for IE.
 * Added `object-fit` property in css.
+* Added prop imageFit.
 
 3.10.0 - (June 28, 2019)
 ------------------