+++ packages/terra-image/CHANGELOG.md
@@ -8,6 +8,11 @@ Unreleased
 * Added `object-fit` property in css.
 * Added prop imageFit.
 
+3.11.0  - (July 11, 2019)
+------------------
+### Changed
+* Minor dependency version bump
+
 3.10.0 - (June 28, 2019)
 ------------------
 ### Changed